Cleaned Image route
Exported module
Shifted to corresponding routes
