+++ routes/images.js
@@ -1,3 +1,9 @@
+const crypto = require('crypto')
+const path = require('path')
+const multer = require('multer')
+const GridFsStorage = require('multer-gridfs-storage')
+
+const uri = 'mongodb+srv://heads:heads@cluster0-v6kuo.mongodb.net/techsite?retryWrites=true&w=majority'
 
 const storage = new GridFsStorage({
   url: uri,
@@ -23,78 +29,80 @@ const upload = multer({ storage })
 // <form method="POST" action="/users/profile/upload/<%=id%>"
 // invoked from form to upload
 
-app.post('/upload/:id', upload.single('file'), (req, res) => {
-  const id = req.params.id
-  res.redirect(`/users/profile/image/update?id=${id}&?url=${req.file.filename}`)
-})
+// app.post('/upload/:id', upload.single('file'), (req, res) => {
+//   const id = req.params.id
+//   res.redirect(`/users/profile/image/update?id=${id}&?url=${req.file.filename}`)
+// })
 
-// returns an image stream to show as prof pic    || todo add it to the ejs once the server is made online
-app.get('/:filename', (req, res) => {
-  const file = gfs
-    .find({
-      filename: req.params.filename
-    })
-    .toArray((err, files) => {
-      if (!files || files.length === 0) {
-        return res.status(404).json({
-          err: 'no files exist'
-        })
-      }
-      gfs.openDownloadStreamByName(req.params.filename).pipe(res)
-    })
-})
+// // returns an image stream to show as prof pic    || todo add it to the ejs once the server is made online
+// app.get('/:filename', (req, res) => {
+//   const file = gfs
+//     .find({
+//       filename: req.params.filename
+//     })
+//     .toArray((err, files) => {
+//       if (!files || files.length === 0) {
+//         return res.status(404).json({
+//           err: 'no files exist'
+//         })
+//       }
+//       gfs.openDownloadStreamByName(req.params.filename).pipe(res)
+//     })
+// })
 
-app.post('/delete/:img', (req, res) => {
-  gfs.delete(new mongoose.Types.ObjectId(req.params.img), (err, data) => {
-    if (err) return res.status(404).json({ err: err.message })
-    res.status(200)
-  })
-})
+// app.post('/delete/:img', (req, res) => {
+//   gfs.delete(new mongoose.Types.ObjectId(req.params.img), (err, data) => {
+//     if (err) return res.status(404).json({ err: err.message })
+//     res.status(200)
+//   })
+// })
 
-// to save the event in database
-// <form action="/users/add_event/<%=club_head_id%>/add_event/<%=club_name%>/add" method="POST" enctype="multipart/form-data">
-// from the add_event form
-app.post('/add_event', upload.single('poster'), (req, res) => {
-  const event = new Event({
-    name: req.body.event_name,
-    venue: req.body.event_venue,
-    date: req.body.event_date,
-    description: req.body.description,
-    poster_url: `/events/posters/${req.file.filename}`,
-    owner: new mongoose.Types.ObjectId(req.body.club_head_id),
-    categories: req.body.categories,
-    speaker: req.body.speaker
-  })
-  event.save((err) => { // saving the event in database
-    console.error.bind(console, 'event not saved check the input')
-  })
+// // to save the event in database
+// // <form action="/users/add_event/<%=club_head_id%>/add_event/<%=club_name%>/add" method="POST" enctype="multipart/form-data">
+// // from the add_event form
+// app.post('/add_event', upload.single('poster'), (req, res) => {
+//   const event = new Event({
+//     name: req.body.event_name,
+//     venue: req.body.event_venue,
+//     date: req.body.event_date,
+//     description: req.body.description,
+//     poster_url: `/events/posters/${req.file.filename}`,
+//     owner: new mongoose.Types.ObjectId(req.body.club_head_id),
+//     categories: req.body.categories,
+//     speaker: req.body.speaker
+//   })
+//   event.save((err) => { // saving the event in database
+//     console.error.bind(console, 'event not saved check the input')
+//   })
 
-  // poupulate the owner field
-  Event.findById(event._id)
-    .populate({
-      path: 'owner',
-      model: 'Users'
-    })
-    .exec((err, event) => {
-      console.error.bind(console, ' error while executing the event population')
-    })
+//   // poupulate the owner field
+//   Event.findById(event._id)
+//     .populate({
+//       path: 'owner',
+//       model: 'Users'
+//     })
+//     .exec((err, event) => {
+//       console.error.bind(console, ' error while executing the event population')
+//     })
 
-  res.send('success')
-})
+//   res.send('success')
+// })
 
-app.get('/:poster', (req, res) => {
-  const filename = req.params.poster
+// app.get('/:poster', (req, res) => {
+//   const filename = req.params.poster
 
-  const file = gfs
-    .find({
-      filename: req.params.poster
-    })
-    .toArray((err, files) => {
-      if (!files || files.length === 0) {
-        return res.status(404).json({
-          err: 'no files exist'
-        })
-      }
-      gfs.openDownloadStreamByName(req.params.poster).pipe(res)
-    })
-})
+//   const file = gfs
+//     .find({
+//       filename: req.params.poster
+//     })
+//     .toArray((err, files) => {
+//       if (!files || files.length === 0) {
+//         return res.status(404).json({
+//           err: 'no files exist'
+//         })
+//       }
+//       gfs.openDownloadStreamByName(req.params.poster).pipe(res)
+//     })
+// })
+
+module.exports = upload ;
