+++ routes/users.js
@@ -2,6 +2,7 @@ const router = require('express').Router()
 const Users = require('../models/Users')
 const Events = require('../models/Event')
 const moment = require('moment')
+var upload = require('./images');
 
 var sess
 
@@ -67,6 +68,56 @@ router.route('/profile/').get((req, res) => {
       res.json(err)
     })
 })
+// <form method="POST" action="/users/profile/image/upload/<%=id%>"
+// invoked from form to upload
+router.route('/profile/image/upload/:id').post( upload.single('file'), (req, res) => {
+  const id = req.params.id
+  // Sending back file name to server
+  // console.log(req.file);
+  res.redirect(`/users/profile/image/update?id=${id}&url=${req.file.filename}`)
+  // res.json({file:req.file});
+})
+
+// use the following format for update requests in this route
+router.route('/profile/').post( upload.single('profpic'), function (req, res, next) {
+  var sess = req.session
+  const id = req.session._id
+  const uid = req.params.id
+  var dpurl = req.body.dp_url
+
+  if (req.file != undefined) {
+    dpurl = req.file.filename
+  }
+  const change = {
+    // pswd: req.body.pswd,
+    dp_url: dpurl,
+    name: req.body.name,
+    contact: req.body.contact,
+    email_id: req.body.email_id,
+    bio: req.body.bio
+  }
+  sess.dp_url = dpurl
+  sess.email_id = req.body.email_id
+  sess.contact = req.body.contact
+  sess.name = req.body.name
+  sess.bio = req.body.bio
+  Users.findByIdAndUpdate(id, change)
+    .then(() => {
+      res.render('public_landing', {
+        id: sess._id,
+        club_name: sess.club_name,
+        name: sess.name,
+        user_id: sess.user_id,
+        pswd: sess.pswd,
+        email_id: sess.email_id,
+        contact: sess.contact,
+        bio: sess.bio,
+        dp_url: sess.dp_url
+      })
+    }).catch(err => {
+      res.json(err)
+    })
+})
 
 // primary route to add event
 router.route('/add_event/').get((req, res) => {
@@ -103,6 +154,39 @@ router.route('/events/details/:id').get((req, res) => {
     }).catch((err) => {
       res.json('Error: ' + err)
     })
+})
+
+// to save the event in database
+// <form action="/users/add_event/<%=club_head_id%>/add_event/<%=club_name%>/add" method="POST" enctype="multipart/form-data">
+// from the add_event form
+router.route('/add_event/save').post( upload.single('poster'), (req, res) => {
+  let poster_url
+  if (req.file == undefined) {
+    poster_url = ' '
+  } else {
+    poster_url = `${req.file.filename}`
+  }
+
+  const event = new Events({
+    name: req.body.event_name + '',
+    venue: req.body.event_venue,
+    date: req.body.event_date,
+    description: req.body.description,
+    poster_url: poster_url, // url to find poster of the event
+    owner: req.session._id,
+    categories: req.body.categories,
+    speaker: req.body.speaker
+
+  })
+
+  event.save((err, event) => { // saving the event in database
+    if (err) {
+      res.json(err)
+    } else {
+      res.redirect('/users/events/retrieve/')
+    }
+  })
+  // let headid = req.params.club_head_id;
 })
 
 // route for using in the notify section
